examples/type-adapter/pyproject.toml
[project]
name = "type-adapter"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "sqlalchemy-pydantic-type",
]

[tool.uv.sources]
